requirements.txt
sanic==23.6.0
Sanic-Cors==2.2.0
sanic-ext==23.6.0
sanic-routing==23.6.0
sanic-testing==23.6.0
sanic-redis==0.3.0
termcolor==1.1.0
python-dotenv==0.19.2
jsonschema==3.2.0
pymongo==3.11.4
click==8.0.3
PyJWT==2.8.0
pydantic==2.4.2
python-Levenshtein==0.26.1
haversine==2.9.0
sentence_transformers==3.3.1